More work on JPEG stuff, think the embedding algorithm is now working, trying to fix memory corruption issues...

+++ src/video/avi_decoder.cc
@@ -27,7 +27,7 @@ class AviChunkWrapper : public Chunk {
     AviChunk *c;
   public:
     AviChunkWrapper(AviChunk *c): c(c) {};
-    virtual int getChunkSize() {
+    virtual long getChunkSize() {
       return c->chunkSize;
     };
     virtual char *getFrameData() {
